评论列表展示实现

+++ src/main/resources/static/js/imustBBS.js
@@ -2,33 +2,40 @@ function comment() {
     let postId = $("#post_id").val();
     let content = $("#comment_content").val();
 
-    $.ajax({
-        type: "POST",
-        url: "/own/comment",
-        contentType: "application/json;charset=utf-8",
-        data: JSON.stringify(
-            {
-                "parentId": postId,
-                "content": content,
-                "type": 1
-            }
-        ),
-        success: function (response) {
-            if (response.code == 200) {
-                $("#comment_section").hide();
-            } else {
-                if (response.code == 2000) {
-                    let isAccepted = confirm(response.message);
-                    if (isAccepted) {
-                        window.open("/login/github");
-
-                        window.localStorage.setItem("closeable", true);
-                    }
-                } else {
-                    alert(response.message);
+    if ($('#comment_content').val().trim().length == 0){
+        alert("评论不能为空");
+    }else {
+        $.ajax({
+            type: "POST",
+            url: "/own/comment",
+            contentType: "application/json;charset=utf-8",
+            data: JSON.stringify(
+                {
+                    "parentId": postId,
+                    "content": content,
+                    "type": 1
                 }
-            }
-        },
-        dataType: "json"
-    });
+            ),
+            beforeSend: function(){
+
+            },
+            success: function (response) {
+                if (response.code == 200) {
+                    $("#comment_section").hide();
+                } else {
+                    if (response.code == 2000) {
+                        let isAccepted = confirm(response.message);
+                        if (isAccepted) {
+                            window.open("/login/github");
+
+                            window.localStorage.setItem("closeable", true);
+                        }
+                    } else {
+                        alert(response.message);
+                    }
+                }
+            },
+            dataType: "json"
+        });
+    }
 }
